+++ tw-tasks-core/src/main/java/com/transferwise/tasks/IEnvironmentValidator.java
@@ -1,0 +1,5 @@
+package com.transferwise.tasks;
+
+public interface IEnvironmentValidator {
+  void validate();
+}