+++ tw-tasks-core/src/main/java/com/transferwise/tasks/IMigrationHandler.java
@@ -1,5 +1,0 @@
-package com.transferwise.tasks;
-
-public interface IMigrationHandler {
-  void migrate();
-}