+++ tw-tasks-core/src/main/java/com/transferwise/tasks/TasksService.java
@@ -57,7 +57,7 @@ public class TasksService implements ITasksService, GracefulShutdownStrategy {
   @Autowired
   private IEntryPointsService entryPointsHelper;
   @Autowired
-  private IMigrationHandler migrationHandler;
+  private IEnvironmentValidator migrationHandler;
 
   private ExecutorService afterCommitExecutorService;
   private TxSyncAdapterFactory txSyncAdapterFactory;
@@ -67,7 +67,7 @@ public class TasksService implements ITasksService, GracefulShutdownStrategy {
 
   @PostConstruct
   public void init() {
-    migrationHandler.migrate();
+    migrationHandler.validate();
 
     if (tasksProperties.isAsyncTaskTriggering()) {
       afterCommitExecutorService = executorsHelper.newScheduledExecutorService("tsac", tasksProperties.getAsyncTaskTriggeringsConcurrency());